Implement basic co-inform user authentication
When the dashboard has attribute `coinform` (with value true), a login
button is shown as part of the dashLoader. Authentication is handled
by the co-inform gateway, which required some tweaking of the app.js
responseInterceptors to treat the api.coinform.eu/login correctly.

Added a requirement for jwt_decode module, needed to correctly parse
the token sent back by the gateway.

Main functionality is implemented in the new service `ciUser`, which
exposes method login and defines properties `loggedIn`, `email` and
`token`.