Make it possible to show row headers even when collapsed
In standard Banana, rows are either collapsed (only header is shown)
or expanded (components in row are shown, but header is hidden). It
was suggested that it could be clearer to the user if headers were
always shown. This means rows will occupy valuable screen real-estate,
but we implemented this as an experimental functionality.

We added a new property to row: `showHeader`. The header is now shown
depending on the value of this property (instead of relying solely on
property `collapse`). For now `showHeader` is set to the same value as
`collapse` so that it behaves the same way as before. However, if the
dashboard has property `coinform_experimental`, this value is
hardcoded to `true` so that headers are always shown.

Another new feature is that rows can have a property `showFirstTime`,
which allows such fields to be expanded the first time a user visits a
specific dashboard(version). This is useful for the "about" row, which
is interesting the first time, but not after that. This is implemented
using the `window.localStorage` functionality.