Covid-19 dashboard with experimental features based on feedback
Copy of default dashboard with a long list of improvements suggested
by the WP5 team. Since some of them are experimental, this is a
temporary dashboard to let the WP5 members to verify that they like
the new implementation. Eventually this version of the dashboard will
replace the default one.