Fix sidebar css
The current implementation was causing issues as part of the sidebar
was hidden under the footer. In the end, I moved the footer inside the
main so that it also shifts along with the main area when the sidebar
is visible. This way the sidebar has the full height to its
disposal. Had to add remove standard padding in main as this meant the
footer was no longer border to border. Added class row-container for
the main row container to re-introduce the padding only for the main
rows in the dashboard.