Added active filter UI within the filters panel
This shows a list of currently active filters (regardless of whether
set within the facets panel). We already showed a count of active
filters, but it wasn't kept in sync with filters set in other
panels (e.g. heatmap or terms). Now it show all "terms" filters and we
added a list of the defined filters with an option to remove them.

This means the "remove" buttons no longer need to appear within the
accordeon lists, which wasn't working correctly anyway as it showed
the "remove" option for all the options for the field, regardless of
the value. So having the new "active filters" list fixes this issue.

This required several new css classes for the activeFilter(s), now
defined in the ci theme.  The ci theme has a few extra tweaks:

 (i) tooltip now opacity 90 as 80 made it hard to read in some
 situations where the background already was dark.

 (ii) the tooltip window now has a border, again to help to separate
 it from the background