Show factchecker reviews instead of acred reviews when available
In Co-inform Document view (e.g. used inside of the table panel), we
typically show a tweet next to a credibility review for that tweet. In
some cases, we may actually have a factchecker review for the
tweet; for example, when a user has requested a factcheck of a tweet
and we have received such a factcheck. In those cases it's better to
show the human rating as it's more likely to be correct than our
AI-based analysis.

Implemented by adding a service call in solrSrv; this fetches any such
reviews from the backend. The ciDocument then takes care of setting
the right flags and the template takes shows the right card depending
on the controller's state.